refactor: collapse chat contact repo access
